+++ python/apps/keep_testing/util/dir_watcher.py
@@ -10,14 +10,14 @@ class DirWatcher:  # pylint: disable=too-few-public-methods
     """Watches directories for change events"""
 
     EVENTS_WATCHED = (
-        inotify.constants.IN_MODIFY |
-        inotify.constants.IN_ATTRIB |
-        inotify.constants.IN_MOVED_FROM |
-        inotify.constants.IN_MOVED_TO |
-        inotify.constants.IN_CREATE |
-        inotify.constants.IN_DELETE |
-        inotify.constants.IN_DELETE_SELF |
-        inotify.constants.IN_MOVE_SELF
+        inotify.constants.IN_MODIFY
+        | inotify.constants.IN_ATTRIB
+        | inotify.constants.IN_MOVED_FROM
+        | inotify.constants.IN_MOVED_TO
+        | inotify.constants.IN_CREATE
+        | inotify.constants.IN_DELETE
+        | inotify.constants.IN_DELETE_SELF
+        | inotify.constants.IN_MOVE_SELF
     )
 
     def __init__(self, files: List[str], dirs: List[str]) -> None:
@@ -30,13 +30,13 @@ class DirWatcher:  # pylint: disable=too-few-public-methods
         unique_dirs = DirWatcher._unify_dirs(files, dirs)
 
         self._watcher = inotify.adapters.InotifyTrees(
-            unique_dirs, mask=self.EVENTS_WATCHED)
+            unique_dirs, mask=self.EVENTS_WATCHED
+        )
 
     @staticmethod
     def _unify_dirs(files: List[str], dirs: List[str]) -> List[str]:
         """Removes directories that are subdirs of other directories in the set"""
-        unique_dirs = {os.path.dirname(file)
-                       for file in files} if files else set()
+        unique_dirs = {os.path.dirname(file) for file in files} if files else set()
         unique_dirs = unique_dirs.union(set(dirs) if dirs else set())
         watched: list[str] = []
         for candidate in sorted(unique_dirs):