+++ python/tests/__main__.py
@@ -7,7 +7,7 @@ import unittest
 dir_tests = os.path.dirname(os.path.realpath(__file__))
 
 test_loader = unittest.defaultTestLoader.discover(dir_tests)
-test_runner = unittest.TextTestRunner()
+test_runner = unittest.TextTestRunner(verbosity=2)
 result = test_runner.run(test_loader)
 if not result.wasSuccessful():
     sys.exit(1)