+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## Version 0.2.0
+
+* `keep_testing` now supports TOML configuration files.
+* `keep_testing` now supports single execution mode (`-1`).
+
 ## Version 0.1.0
 
 `keep_testing` is available: monitors directories and files and `ENTER` key.